Internal change
PiperOrigin-RevId: 277794396
