--- README.md
@@ -1,10 +1,0 @@
-# Applied Computing
-
-The [Applied Computing Series](https://edu.google.com/computer-science/applied-computing-series/)
-is a collection of classes focused on teaching data science and machine learning concepts to
-students.
-
-The series is divided into two main parts:
-
-  * [Applied Computing](./acs/)
-  * [Applied Machine Learning Intensive](./amli/)
