--- amli/amli/concepts.yaml
@@ -1,87 +1,0 @@
-- data
-  - analysis
-  - processing
-    - normalization
-    - standardization
-    - feature engineering
-      - singular value decomposition
-      - principal component analysis
-      - embeddings
-- linear algebra
-- toolkits
-  - scikit-learn
-    - random forest classifier
-    - random forest regressor
-  - tensorflow
-- models
-  - tuning
-  - types
-    - classification
-    - regression
-  - measurement
-    - regression
-      - mean absolute error
-      - mean squared error
-      - root mean squared error
-    - classification
-      - accuracy
-      - f1
-      - precision
-      - recall
-      - softmax
-      - homogeneity
-  - techniques
-    - ensemble learning
-    - bootstrapping
-    - bagging
-    - boosting
-  - algorithms
-    - linear regression
-    - polynomial regression
-    - logistic regression
-    - support vector machines
-    - decision trees
-      - gini
-      - entrophy
-    - random forests
-    - k-nearest neighbors
-    - bayesian
-    - perceptrons
-    - deep nerual networks
-    - convolutional neural networks
-    - k-means
-    - recurrent nerual networks
-    - natural language processing
-    - generative adversarial networks
-    - reinforcement learning
-    - transfer learning
-    - recommendation systems
-    - sequence prediction
-- programming
-  - python
-    - data structures
-      - dictionaries
-      - lists
-      - tuples
-      - dqueues
-    - functions
-    - classes
-    - list-expansions
-  - sql
-    - select
-    - filter
-    - inner-join
-    - outer-join
-  - algorithms
-    - binary trees
-- statistics
-  - distributions
-    - normal
-    - non-normal
-- visualizations
-  - boxplot
-  - histogram
-  - calibration plot
-  - correlation matrix
-  - roc curve
-  - scatterplot
