+++ amli/content/t05_basic_classification/04_random_forests/slides.md
@@ -4,9 +4,6 @@
 In this unit we are going to talk about random forests. Random forests build upon a concept that we
 have already learned, decision trees, so let's take a moment to remind ourselves about decision
 trees.
-
-Concepts:
-* models:random forests
 -->
 
 ---
@@ -21,9 +18,6 @@ about Decision Trees.
 
 Image Details:
 * [tree.png](https://pixabay.com/vectors/tree-silhouette-winter-plant-3979965/): Pixabay License
-
-Concepts:
-* models:decision trees
 -->
 
 ---
@@ -44,9 +38,6 @@ the node and then following a path based on the comparison. The leaf node that y
 
 Image Details:
 * tree.png: Copyright Google
-
-Concepts:
-* models:decision trees
 -->
 
 ---
@@ -74,12 +65,6 @@ numeric confidence in the applicability of each label. Instead you receive a sin
 Decision trees are not sensitive to features having different ranges of values. Most machine
 learning algorithms benefit from scaling and/or normalizing data, but decision trees are an
 exception.
-
-Concepts:
-* models:decision trees
-* programming:algorithms:binary trees
-* models:types:classification
-* models:types:regression
 -->
 
 ---
@@ -100,10 +85,6 @@ Entropy
 Also remember that when building decision trees we must choose what factor to base our splits on. We
 can use the Gini algorithm to minimize impurity or entropy to maximize information gain. We
 discussed this in our unit about decision trees.
-
-Concepts:
-* models:algorithms:decision trees:gini
-* models:algorithms:decision trees:entrophy
 -->
 
 ---
@@ -121,9 +102,6 @@ decision made by the random forest.
 
 Image Details:
 * [forest.png](https://pixabay.com/vectors/deciduous-trees-forest-trees-154168/): Pixabay License
-
-Concepts:
-* models:algorithms:random forests
 -->
 
 ---
@@ -134,9 +112,6 @@ Concepts:
 This is our first foray into "ensemble learning". Ensemble learning is a technique in which multiple
 learners are trained on the training data and their results are aggregated in some way. This
 aggregation is typically majority vote, mode, for classification and mean or median for regression.
-
-Concepts:
-* models:techniques:ensemble learning
 -->
 
 ---
@@ -154,9 +129,6 @@ This is considered sampling "with replacement".
 
 You can also choose to train every tree with the entire dataset. In this case you get variation in
 trees based on their random starting points.
-
-Concepts:
-* models:techniques:bootstrapping
 -->
 
 ---
@@ -177,9 +149,6 @@ there will likely be duplicates in each generated dataset. This allows each tree
 different view of the data.
 
 Note that after the datasets are made, trees can be built and used in parallel.
-
-Concepts:
-* models:techniques:bootstrapping
 -->
 
 ---
@@ -195,9 +164,6 @@ predictions that the model before it got wrong.
 You can think of it as an assembly line where each worker has a specialty.
 
 The downside of boosting is that it has to be done sequentially.
-
-Concepts:
-* models:techniques:boosting
 -->
 
 ---
@@ -211,12 +177,6 @@ Concepts:
 scikit-learn provides a couple of random forest implementations, one for classification and one
 for regression. The two implementations share many hyperparameters, but not all. We'll spend the
 next few slides highlighting some of the hyperparameters.
-
-Concepts:
-* toolkits:scikit-learn:random forest classifier
-* toolkits:scikit-learn:random forest regressor
-* models:types:classification
-* models:types:regression
 -->
 
 ---
@@ -242,9 +202,6 @@ How to you choose the best number?
 
 Experimentation. Play with different settings and compare training time and model scores until you
 find a value that seems to be fast and "good" enough for your use case.
-
-Concepts:
-* models:tuning
 -->
 
 ---
@@ -268,12 +225,6 @@ our model quality discussions that MSE penalizes outliers much more than MAE.
 
 For classification we can choose to either use Gini, which is an impurity measurement, or Entropy,
 which is a measure of information gain. Both of these are measures of set homogeneity.
-
-Concepts:
-* models:tuning
-* models:measurement:regression:mean absolute error
-* models:measurement:regression:mean squared error
-* models:measurement:classification:homogenity
 -->
 
 ---
@@ -295,9 +246,6 @@ leaves.
 
 Let's look at another hyperparameter that can help prevent overfitting without having to worry
 about balance.
-
-Concepts:
-* models:tuning
 -->
 
 ---
@@ -323,9 +271,6 @@ The nice thing about these hyperparameters are that they work well with unbalanc
 `max_depth`.
 
 How do you pick a value for these parameters? Trial and error, like most hyperparameters.
-
-Concepts:
-* models:tuning
 -->
 
 ---
@@ -339,9 +284,6 @@ By default a decision tree in a random forest will attempt to use all features i
 best split for a node. If you have a lot of features this might be computationally expensive or
 even lead to some overfitting. You can limit the number of features used in any decision so that
 only the most significant features are used at each node.
-
-Concepts:
-* models:tuning
 -->
 
 ---
@@ -359,7 +301,4 @@ more exist. Note that there are some differences in what parameters you can twea
 and classifiers. Also, any parameters you set apply to all trees in a forest. And finally, the best
 way to find the best parameters is to experiment with many different options and find which perform
 best for your data.
-
-Concepts:
-* models:tuning
--->
+!-->
