+++ amli/requirements.txt
@@ -2,4 +2,3 @@ absl-py
 google-auth-oauthlib
 nbformat
 pytest
-pyyaml
